requirements.txt
beautifulsoup4
streamlit
anthropic
google-generativeai
python-dotenv
pandas
requests
